android_app/requirements.txt
kivy>=2.1.0
reportlab>=3.6.0
Pillow>=9.0.0
pandas>=1.3.0
openpyxl>=3.0.0
buildozer>=1.4.0
